try correlations again
backprop requires derivating a gradient against loss, but there's no
clear equation between loss (difference of target brightness to actual
brightness of edges map) and the obtained thresholds (that's why we're
doing this in the first place)

if there's one, it could be found by a simple grid search.

Trying with brightness and contrast for now, trying to detect edges on
cropped segments that contain visible fracture lines.

Marking settings that yield a satisfying result and comparing quotient
and difference of brigthness of edgemap to original image (the target
brightness might actually depend on the original image's?)

also brightness of edgemap depends on the number of actual segments
in the image, which is linked to contrast, so finding a link between
original image's contrast and brightness of edgemap might be worth it?